refactor(parse): move unbacked-grid-char check into validation module
The rebus-glyph warning check and its tests lived in parser/mod.rs; move them to
validation.rs alongside validate_puzzle, where puzzle validation belongs. The
function stays warning-producing (validate_puzzle remains error-only), so mod.rs
still extends its warnings with the result. No behavior change.
