+++ parse/src/parser/validation.rs
@@ -1,6 +1,6 @@
 use crate::{
-    error::PuzError,
-    grid::{count_clues, TAKEN_SQUARE},
+    error::{PuzError, PuzWarning},
+    grid::{count_clues, is_standard_cell_char, TAKEN_SQUARE},
     types::Puzzle,
 };
 
@@ -9,6 +9,41 @@ pub(crate) fn validate_puzzle(puzzle: &Puzzle) -> Result<(), PuzError> {
     validate_grid_structure(&puzzle.grid.blank, &puzzle.grid.solution)?;
     validate_clue_consistency(puzzle)?;
     Ok(())
+}
+
+/// Warn about solution cells that hold a non-standard character with no rebus
+/// entry backing them.
+///
+/// A `.puz` file places no constraint on cell bytes, and rebus puzzles put
+/// arbitrary glyphs (`#`, `*`, high bytes, and so on) in solution cells. Those
+/// are valid when the GRBS grid marks the cell as a rebus. A non-standard char
+/// with no such backing is unusual and may indicate corruption, so we warn
+/// rather than reject the file. Unlike [`validate_puzzle`], this produces
+/// warnings, not errors.
+pub(crate) fn check_unbacked_grid_chars(puzzle: &Puzzle) -> Vec<PuzWarning> {
+    let mut warnings = Vec::new();
+    let rebus_grid = puzzle.extensions.rebus.as_ref().map(|r| &r.grid);
+
+    for (row, line) in puzzle.grid.solution.iter().enumerate() {
+        for (col, ch) in line.chars().enumerate() {
+            if is_standard_cell_char(ch) {
+                continue;
+            }
+            let backed = rebus_grid
+                .and_then(|g| g.get(row))
+                .and_then(|r| r.get(col))
+                .is_some_and(|&key| key != 0);
+            if !backed {
+                warnings.push(PuzWarning::UnbackedGridChar {
+                    character: ch,
+                    row,
+                    col,
+                });
+            }
+        }
+    }
+
+    warnings
 }
 
 fn validate_puzzle_dimensions(width: u8, height: u8) -> Result<(), PuzError> {
@@ -84,7 +119,8 @@ fn validate_clue_consistency(puzzle: &Puzzle) -> Result<(), PuzError> {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::types::{Clues, Extensions, Grid, Puzzle, PuzzleInfo};
+    use crate::error::PuzWarning;
+    use crate::types::{Clues, Extensions, Grid, Puzzle, PuzzleInfo, Rebus};
     use std::collections::HashMap;
 
     /// Helper to create a valid test puzzle
@@ -300,5 +336,92 @@ mod tests {
         } else {
             panic!("Expected InvalidDimensions error");
         }
+    }
+
+    // --- check_unbacked_grid_chars ---
+
+    /// Build a puzzle from a solution grid, deriving the blank grid and taking
+    /// an optional rebus extension.
+    fn puzzle_with_solution(solution: Vec<String>, rebus: Option<Rebus>) -> Puzzle {
+        let width = solution[0].chars().count() as u8;
+        let height = solution.len() as u8;
+        let blank = solution
+            .iter()
+            .map(|row| {
+                row.chars()
+                    .map(|c| if c == '.' { '.' } else { '-' })
+                    .collect()
+            })
+            .collect();
+        Puzzle {
+            info: PuzzleInfo {
+                title: String::new(),
+                author: String::new(),
+                copyright: String::new(),
+                notes: String::new(),
+                width,
+                height,
+                version: "1.3".to_string(),
+                is_scrambled: false,
+            },
+            grid: Grid { blank, solution },
+            clues: Clues {
+                across: HashMap::new(),
+                down: HashMap::new(),
+            },
+            extensions: Extensions {
+                rebus,
+                circles: None,
+                given: None,
+            },
+        }
+    }
+
+    #[test]
+    fn test_plain_grid_produces_no_warning() {
+        let p = puzzle_with_solution(vec!["AB".into(), "CD".into()], None);
+        assert!(check_unbacked_grid_chars(&p).is_empty());
+    }
+
+    #[test]
+    fn test_unbacked_marker_char_warns() {
+        // '#' at (0,0) with no rebus data.
+        let p = puzzle_with_solution(vec!["#B".into(), "CD".into()], None);
+        let warnings = check_unbacked_grid_chars(&p);
+        assert_eq!(warnings.len(), 1);
+        assert!(matches!(
+            warnings[0],
+            PuzWarning::UnbackedGridChar {
+                character: '#',
+                row: 0,
+                col: 0
+            }
+        ));
+    }
+
+    #[test]
+    fn test_marker_char_backed_by_rebus_is_silent() {
+        // '#' at (0,0), and the rebus grid marks that cell.
+        let mut table = HashMap::new();
+        table.insert(1u8, "HASH".to_string());
+        let rebus = Rebus {
+            grid: vec![vec![1, 0], vec![0, 0]],
+            table,
+        };
+        let p = puzzle_with_solution(vec!["#B".into(), "CD".into()], Some(rebus));
+        assert!(check_unbacked_grid_chars(&p).is_empty());
+    }
+
+    #[test]
+    fn test_high_byte_char_backed_by_rebus_is_silent() {
+        // 'Â' (0xC2) at (0,0) backed by a rebus entry.
+        let mut table = HashMap::new();
+        table.insert(1u8, "CENT".to_string());
+        let rebus = Rebus {
+            grid: vec![vec![1, 0], vec![0, 0]],
+            table,
+        };
+        let p = puzzle_with_solution(vec!["\u{00C2}B".into(), "CD".into()], Some(rebus));
+        assert!(check_unbacked_grid_chars(&p).is_empty());
     }
 }
